use tableInfo.json fix issue with nullable text for formatters remove kotlin reflect

+++ qmobileui/src/main/java/com/qmobile/qmobileui/action/utils/DotProgressBar.kt
@@ -83,7 +83,7 @@ class DotProgressBar : FrameLayout {
         }
         primaryAnimator?.repeatMode = ValueAnimator.RESTART
         primaryAnimator?.repeatCount = ValueAnimator.INFINITE
-        primaryAnimator?.duration = ANIMATION_DURATION.toLong()
+        primaryAnimator?.duration = ANIMATION_DURATION
         primaryAnimator?.interpolator = LinearInterpolator()
 
         startAnimation()
